Version 3 | Hash in Python
Created the initial draft of the Hash.md file. This file is ballooning in line numbers, it's getting hard to keep track of. I think I just want to push the file to the repo, read it there outside of the preview and figure the rest of the file out from there.

Also, gave up on managing the versioning. We'll see if I end up regretting that decision in the future.