Version 4.0 | Introducing Markdown.md
Decided to create documentation for all the learning I've done in Markdown. I bet I'm missing a couple things that I already know at this time - but it's 11 PM and I still haven't ate yet and I have work in the morning.

So...

¯\_(ツ)_/¯

(Oh, also... finished Hash.md I guess. I'm so bored of the file, I'm certainly done with it for now)